requirements.txt
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
streamlit==1.28.1
pillow==10.0.1
scikit-learn==1.3.0
seaborn==0.12.2
plotly==5.17.0
PyMySQL==1.1.0
openpyxl==3.1.5
